+++ client/src/App.tsx
@@ -7,7 +7,7 @@ import { addZetel } from "./redux/modules/zettelReducer";
 import { getZettels } from "./api/zettelApi";
 import LoginCallbackPage from "./pages/LoginCallbackPage";
 import useCurrentUser from "./hooks/useCurrentUser";
-import Toast from "./components/Toast";
+import Toast from "./components/core/Toast";
 
 function App() {
   const dispatch = useDispatch();
