+++ client/src/components/__test__/Tag.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent } from "@testing-library/react";
-import Tag from "../Tag";
+import Tag from "../common/Tag";
 
 describe("<Tag />", () => {
   it("renders properly", () => {
